Print instead of logging during config parseing.
