+++ sydent/sydent.py
@@ -22,8 +22,9 @@ from typing import Optional
 
 import twisted.internet.reactor
 from twisted.internet import address, task
+from twisted.python import log
 
-from sydent.config import SydentConfig, setup_logging_from_file
+from sydent.config import SydentConfig
 from sydent.db.hashing_metadata import HashingMetadataStore
 from sydent.db.sqlitedb import SqliteDatabase
 from sydent.db.valsession import ThreePidValSessionStore
@@ -306,12 +307,45 @@ def run_gc():
             gc.collect(i)
 
 
-if __name__ == "__main__":
-    config_file = get_config_file_path()
-    setup_logging_from_file(config_file)
+def setup_logging(config: SydentConfig) -> None:
+    """
+    Setup logging using the options specified in the config
 
+    :param config: the configuration to use
+    """
+    log_path = config.general.log_path
+    log_level = config.general.log_level
+
+    log_format = "%(asctime)s - %(name)s - %(lineno)d - %(levelname)s" " - %(message)s"
+    formatter = logging.Formatter(log_format)
+
+    if log_path != "":
+        handler: logging.StreamHandler = logging.handlers.TimedRotatingFileHandler(
+            log_path, when="midnight", backupCount=365
+        )
+        handler.setFormatter(formatter)
+
+        def sighup(signum, stack):
+            logger.info("Closing log file due to SIGHUP")
+            handler.doRollover()
+            logger.info("Opened new log file due to SIGHUP")
+
+    else:
+        handler = logging.StreamHandler()
+
+    handler.setFormatter(formatter)
+    rootLogger = logging.getLogger("")
+    rootLogger.setLevel(log_level)
+    rootLogger.addHandler(handler)
+
+    observer = log.PythonLoggingObserver()
+    observer.start()
+
+
+if __name__ == "__main__":
     sydent_config = SydentConfig()
-    sydent_config.parse_config_file(config_file)
+    sydent_config.parse_config_file(get_config_file_path())
+    setup_logging(sydent_config)
 
     syd = Sydent(sydent_config)
     syd.run()
